Making the inbox update real-time
I started down the road of trying to hook up websockets, but couldn't
get response.stream to stream html. So, I left the wiring in and just
had the "live" response return a dumb string, and in the EventSource
callback, I make an ajax call for the inbox html. I'm explaining all
this because, I could have just started with a long poll, strictly
javascript that makes an ajax call periodically, but instead have this
thing that makes two serve requests. A bit clunky, but it works, so
leaving it in until we can do it the right way at some point.